+++ frontend/src/typings/MetadataOptionsSettings.ts
@@ -1,0 +1,7 @@
+// krzw(imdb-title-provider)
+export default interface MetadataOptionsSettings {
+  imdbTitleProviderEnabled: boolean;
+  imdbTitleProviderRegions: string;
+  imdbTitleProviderLanguages: string;
+  imdbTitleProviderRefreshInterval: number;
+}